# swagger.yaml
openapi: 3.0.0
info:
  title: Express-Portfolio API
  contact: {}
  version: "1.0"
servers:
  - url: https://agliofolio.com
    variables: {}
paths:
...
  /find-password:
    get:
      tags:
        - /
      summary: 비밀번호 찾기 페이지를 렌더합니다
      operationId: getFindPassword
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
    post:
      tags:
        - /
      summary: 비밀번호를 찾습니다
      operationId: findPassword
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindPasswordRequest"
            example:
              email: a@a.com
              authNumber: 123456
        required: true
      responses:
        "200":
          description: 이메일이 입력되면 인증번호를 입력한 이메일로 보냅니다. 인증번호가 입력되면 비밀번호 초기화 페이지를 렌더합니다
          headers: {}
      deprecated: false
  /init-password:
    get:
      tags:
        - /
      summary: 비밀번호 초기화 페이지를 렌더합니다
      operationId: getInitPassword
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
    post:
      tags:
        - /
      summary: 비밀번호를 초기화합니다
      operationId: initPassword
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitPasswordRequest"
            example:
              password: secret
              confirm: secret
        required: true
      responses:
        "200":
          description: ""
          headers: {}
        "400":
          description: 잘못된 요청 "password, confirm"
          headers: {}
      deprecated: false
  /watch/${id(\w{32})}:
    get:
      tags:
        - /
      summary: 특정 비디오를 시청합니다
      operationId: getWatch
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
components:
  schemas:
...
    FindPasswordRequest:
      title: FindPasswordRequest
      required:
        - email
        - authNumber
      properties:
        email:
          type: string
        authNumber:
          type: number
      example:
        email: a@a.com
        authNumber: 123456
    InitPasswordRequest:
      title: InitPasswordRequest
      required:
        - password
        - confirm
      properties:
        password:
          type: string
        confirm:
          type: string
      example:
        password: secret
        confirm: secret
...